requirements.txt
pygame
PyOpenGL
pyassimp
